Create get by id operation
- Refactor code to keep pattern on the usecases
- Refactor RepoImpl to reduce code

+++ src/main/java/com/demo/clean_architecture/infra/db/mysql/repositories/StudentRepository.java
@@ -5,6 +5,6 @@ import com.demo.clean_architecture.infra.db.mysql.entities.StudentDto;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentRepository extends JpaRepository<StudentDto, Integer> {
+public interface StudentRepository extends JpaRepository<StudentDto, Long> {
 
 }
